+++ concordium-consensus/tests/scheduler/Spec.hs
@@ -40,6 +40,7 @@ import qualified SchedulerTests.SmartContracts.V1.Fallback (tests)
 import qualified SchedulerTests.SmartContracts.V1.Checkpointing (tests)
 import qualified SchedulerTests.SmartContracts.V1.AllNewHostFunctions (tests)
 import qualified SchedulerTests.SmartContracts.V1.Upgrading (tests)
+import qualified SchedulerTests.SmartContracts.V1.Queries (tests)
 
 import Test.Hspec
 
@@ -82,5 +83,6 @@ main = hspec $ do
          SchedulerTests.SmartContracts.V1.Checkpointing.tests
          SchedulerTests.SmartContracts.V1.AllNewHostFunctions.tests
          SchedulerTests.SmartContracts.V1.Upgrading.tests
+         SchedulerTests.SmartContracts.V1.Queries.tests
          SchedulerTests.Payday.tests
          SchedulerTests.Delegation.tests